update items for webpack
