+++ package.json
@@ -4,7 +4,8 @@
   "description": "Plan your life using this planner! Includes daily, weekly, monthly calendars that syncs with your chrome storage",
   "main": "webpack.common.js",
   "scripts": {
-    "start": "webpack serve --config webpack.common.js"
+    "start": "webpack serve --config webpack.dev.js",
+    "build": "webpack --config webpack.prod.js"
   },
   "repository": {
     "type": "git",
@@ -20,14 +21,17 @@
   "devDependencies": {
     "clean-webpack-plugin": "*",
     "css-loader": "^5.2.2",
+    "css-minimizer-webpack-plugin": "^2.0.0",
     "file-loader": "^6.2.0",
     "html-loader": "^2.1.2",
     "html-webpack-plugin": "^5.3.1",
+    "json-minimizer-webpack-plugin": "^2.1.0",
     "mini-css-extract-plugin": "^1.4.1",
     "node-sass": "^5.0.0",
     "sass-loader": "^11.0.1",
     "style-loader": "^2.0.0",
     "webpack": "^5.33.2",
+    "webpack-cli": "^4.6.0",
     "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^5.7.3"
   }
