+++ src/index.js
@@ -1,0 +1,3 @@
+import { startApp } from './app/app'
+
+startApp();
