+++ src/manifest.json
