+++ src/template.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <title>Polus</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge" />
+    <!-- <link rel="shortcut icon" type='image/png' href="../images/polus_tab_icon.png" /> -->
+</head>
+
+<body>
+    <ul id="nav">
+        <li id="app-info-box">
+            <div id="app-info">
+                <div id="app-title">
+                    <!-- <img class='app-icon' src="../images/polus_icon.png"></img> -->
+                    <!-- <span class='app-sub'>Polus</span> -->
+                </div>
+                <div id="app-items">
+                    <!-- <div id="settings">Settings</div> -->
+                    <div id="site"><a target="_blank" rel="noopener noreferrer"
+                            href='https://danielchicchon.io/polus'>Site</a>
+                    </div>
+                    <div id="contact">
+                        <a target="_blank" rel="noopener noreferrer"
+                            href="https://docs.google.com/forms/d/e/1FAIpQLSdHcPhbcAWeWFvEFqF6qzmPUD0UtNn9e7pn_eLUukGLudMy1A/viewform">
+                            Feedback</a>
+                    </div>
+                    <div><a id='pmode' target="_blank" rel="noopener noreferrer">Photo Mode</a></div>
+                    <!-- If user not signed in -->
+                    <!-- <div><a id='auth' target="_blank" rel="noopener noreferrer"></a></div> -->
+
+                </div>
+            </div>
+        </li>
+        <li id="background-info-box" style="float: right;">
+            <div id="background-info">
+                <span id="background-location"></span>
+                <span id="background-source">Photo by
+                    <a id="photo-link" target="_blank" rel="noopener noreferrer"></a> on
+                    <a id="site-link" target="_blank" rel="noopener noreferrer"
+                        href="https://unsplash.com/?utm_source=Planner&utm_medium=referral">Unsplash
+                    </a>
+                    <a id='download' rel='nofollow' target='_blank' download>&#8681;</a>
+                </span>
+            </div>
+        </li>
+    </ul>
+
+    <main>
+        <div id="date"></div>
+        <div id="clock"></div>
+        <div id="view">
+            <button type='button' class="view-btn">Daily</button>
+            <button type='button' class="view-btn">Week</button>
+            <button type='button' class="view-btn">Month</button>
+        </div>
+
+        <section id="daily"></section>
+        <section id="week"></section>
+        <section id="month"></section>
+    </main>
+
+    <!-- <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-app.js"></script> -->
+    <!-- <script src="https://www.gstatic.com/firebasejs/8.0.2/firebase-auth.js"></script> -->
+    <!-- <script type="text/javascript" src="../js/auth.js"></script> -->
+    <!-- <script type="text/javascript" src="../js/view.js"></script>
+    <script type="text/javascript" src="../js/clock.js"></script>
+    <script type="text/javascript" src="../js/helper.js"></script>
+    <script type="text/javascript" src="../js/day.js"></script>
+    <script type="text/javascript" src="../js/week.js"></script>
+    <script type="text/javascript" src="../js/month.js"></script>
+    <script type='text/javascript' src='../js/options.js'></script>
+    <script type="text/javascript" src="../js/start.js"></script> -->
+
+</body>
+
+</html>
