+++ webpack.dev.js
@@ -1,0 +1,11 @@
+
+const { merge } = require("webpack-merge")
+const common = require("./webpack.common")
+
+module.exports = merge(common, {
+    mode: "development",
+    output: {
+        filename: "[name].bundle.js",
+        path: path.resolve(__dirname, "dist")
+    }
+})
